+++ ext/duckdb/appender.c
@@ -10,7 +10,7 @@ static VALUE appender_error_message(VALUE self);
 static VALUE appender__append_bool(VALUE self, VALUE val);
 static VALUE appender__append_int8(VALUE self, VALUE val);
 static VALUE appender__apend_int16(VALUE self, VALUE val);
-static VALUE appender_append_int32(VALUE self, VALUE val);
+static VALUE appender__append_int32(VALUE self, VALUE val);
 static VALUE appender_append_int64(VALUE self, VALUE val);
 static VALUE appender_append_uint8(VALUE self, VALUE val);
 static VALUE appender_append_uint16(VALUE self, VALUE val);
@@ -147,32 +147,14 @@ static VALUE appender__apend_int16(VALUE self, VALUE val) {
     return duckdb_state_to_bool_value(duckdb_append_int16(ctx->appender, i16val));
 }
 
-/* call-seq:
- *   appender.append_int32(val) -> self
- *
- * Appends an int32(INTEGER) value to the current row in the appender.
- *
- *   require 'duckdb'
- *   db = DuckDB::Database.open
- *   con = db.connect
- *   con.query('CREATE TABLE users (id INTEGER, age INTEGER)')
- *   appender = con.appender('users')
- *   appender
- *     .append_int32(1)
- *     .append_int32(20)
- *     .end_row
- *     .flush
- */
-static VALUE appender_append_int32(VALUE self, VALUE val) {
+/* :nodoc: */
+static VALUE appender__append_int32(VALUE self, VALUE val) {
     rubyDuckDBAppender *ctx;
     int32_t i32val = (int32_t)NUM2INT(val);
 
     TypedData_Get_Struct(self, rubyDuckDBAppender, &appender_data_type, ctx);
 
-    if (duckdb_append_int32(ctx->appender, i32val) == DuckDBError) {
-        rb_raise(eDuckDBError, "failed to append int32");
-    }
-    return self;
+    return duckdb_state_to_bool_value(duckdb_append_int32(ctx->appender, i32val));
 }
 
 /* call-seq:
@@ -458,7 +440,6 @@ void rbduckdb_init_duckdb_appender(void) {
     rb_define_alloc_func(cDuckDBAppender, allocate);
     rb_define_method(cDuckDBAppender, "initialize", appender_initialize, 3);
     rb_define_method(cDuckDBAppender, "error_message", appender_error_message, 0);
-    rb_define_method(cDuckDBAppender, "append_int32", appender_append_int32, 1);
     rb_define_method(cDuckDBAppender, "append_int64", appender_append_int64, 1);
     rb_define_method(cDuckDBAppender, "append_uint8", appender_append_uint8, 1);
     rb_define_method(cDuckDBAppender, "append_uint16", appender_append_uint16, 1);
@@ -481,6 +462,7 @@ void rbduckdb_init_duckdb_appender(void) {
     rb_define_private_method(cDuckDBAppender, "_append_bool", appender__append_bool, 1);
     rb_define_private_method(cDuckDBAppender, "_append_int8", appender__append_int8, 1);
     rb_define_private_method(cDuckDBAppender, "_append_int16", appender__apend_int16, 1);
+    rb_define_private_method(cDuckDBAppender, "_append_int32", appender__append_int32, 1);
     rb_define_private_method(cDuckDBAppender, "_append_date", appender__append_date, 3);
     rb_define_private_method(cDuckDBAppender, "_append_interval", appender__append_interval, 3);
     rb_define_private_method(cDuckDBAppender, "_append_time", appender__append_time, 4);