+++ lib/duckdb/appender.rb
@@ -160,6 +160,27 @@ module DuckDB
       raise_appender_error('failed to append_int16')
     end
 
+    # call-seq:
+    #   appender.append_int32(val) -> self
+    #
+    # Appends an int32(INTEGER) value to the current row in the appender.
+    #
+    #   require 'duckdb'
+    #   db = DuckDB::Database.open
+    #   con = db.connect
+    #   con.query('CREATE TABLE users (id INTEGER, age INTEGER)')
+    #   appender = con.appender('users')
+    #   appender
+    #     .append_int32(1)
+    #     .append_int32(20)
+    #     .end_row
+    #     .flush
+    def append_int32(value)
+      return self if _append_int32(value)
+
+      raise_appender_error('failed to append_int32')
+    end
+
     # appends huge int value.
     #
     #   require 'duckdb'