+++ lib/duckdb/appender.rb
@@ -202,6 +202,28 @@ module DuckDB
       raise_appender_error('failed to append_int64')
     end
 
+    # call-seq:
+    #  appender.append_uint8(val) -> self
+    #
+    # Appends an uint8 value to the current row in the appender.
+    # The value must be in the range of 0 to 255.
+    #
+    #  require 'duckdb'
+    #  db = DuckDB::Database.open
+    #  con = db.connect
+    #  con.query('CREATE TABLE users (id INTEGER, age UTINYINT)')
+    #  appender = con.appender('users')
+    #  appender
+    #    .append_int32(1)
+    #    .append_uint8(20)
+    #    .end_row
+    #    .flush
+    def append_uint8(value)
+      return self if _append_uint8(value)
+
+      raise_appender_error('failed to append_uint8')
+    end
+
     # appends huge int value.
     #
     #   require 'duckdb'