+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 # Unreleased
 - add 'DuckDB::Appender#error_message'.
 - fix error message when DuckDB::Appender#flush failed.
+- 'DuckDB::Appender#begin_row' does nothing. Only returns self. 'DuckDB::Appender#end_row' is only required.
 
 # 1.1.3.1 - 2024-11-27
 - fix to `DuckDB::Connection#query` with multiple SQL statements. Calling PreparedStatement#destroy after each statement executed.